bugs in the cron scripts.

+++ site/protected/commands/removeTokens.php
@@ -5,7 +5,9 @@
  * in database.
  */
 
+$mysqlhost = null;
 function loadConfig(){
+	global $mysqlhost;
 	return require_once '../config/main.php';
 }
 
@@ -13,6 +15,12 @@ $config = loadConfig();
 
 $db = $config['components']['db'];
 $link = mysql_connect($mysqlhost,$db['username'],$db['password']);
+
+if($link === false){
+	echo "Database connection failed.";
+	throw new Exception("Database connection failed.");
+	exit;
+}
 
 if(!mysql_query('DELETE FROM simpleseoapi.tokens WHERE  expire < UNIX_TIMESTAMP()',$link))
 	echo 'Delete unused tokens failed.'."\nERROR:\n".mysql_errno()."\n\n";
